Add files via upload
//Before using an online service, a user needs to create a valid password. You are to write a program to check a password to verify that 
//it satisfies all five of the following validation criteria:
1. At least one letter in the range 'a'...'z' (inclusive)
2. At least one letter in the rangr 'a'...'z'(inclusive) or at least on of the three characters '@", '#', '$'
3. At least one digit in the range 0...9
4. Minimum length 6 characters
5. Maximum length 15 characters

Input
The input to your program is a single line read from standard input containing no whitespace. The contents of the one line
is the proposed password you are testing

Output
Print "invalid" if the password entered is invalid, otherwise, print "vailid". A password is valif iff it satisfies all five criterias.